Changed player character resource to interface

+++ app/src/app/modules/players-characters/players-characters.component.ts
@@ -22,7 +22,8 @@ export class PlayersCharactersComponent {
 
   playersCharacters = this.#campaignService.playersCharacters;
 
-  deletePlayerCharacter(id: string) {
+  deletePlayerCharacter(id: string | undefined) {
+    if (!id) return;
     this.#campaignService.deletePlayerCharacter(id);
   }
